+++ ntt/include/nncase/ntt/arch/x86_64/arch_types.h
@@ -44,6 +44,10 @@ NTT_END_DEFINE_NATIVE_VECTOR()
 NTT_BEGIN_DEFINE_NATIVE_VECTOR_DEFAULT(uint64_t, __m256i, 4)
 NTT_END_DEFINE_NATIVE_VECTOR()
 
+NTT_DEFINE_NATIVE_VECTOR_DEFAULT_BITCAST(bfloat16, __m256i, __v16hu, uint16_t,
+                                         16)
+NTT_END_DEFINE_NATIVE_VECTOR()
+
 NTT_BEGIN_DEFINE_NATIVE_VECTOR_DEFAULT(float, __m256, 8)
 NTT_END_DEFINE_NATIVE_VECTOR()
