+++ ntt/include/nncase/ntt/native_vector.h
@@ -62,3 +62,26 @@
         static_assert(TIndex::rank() == 1, "index must be 1D");                \
         reinterpret_cast<cast_type &>(array)[index[dim_zero]] = value;         \
     }
+
+#define NTT_DEFINE_NATIVE_VECTOR_DEFAULT_BITCAST(                              \
+    element_type_, native_type, cast_type, bitcast_type, lanes)                \
+    NTT_BEGIN_DEFINE_NATIVE_VECTOR(element_type_, native_type, lanes)          \
+                                                                               \
+    template <Dimensions TIndex>                                               \
+    static element_type_ get_element(const native_type &array,                 \
+                                     const TIndex &index) noexcept {           \
+        static_assert(TIndex::rank() == 1, "index must be 1D");                \
+        const auto casted_array = (cast_type)array;                            \
+        return std::bit_cast<element_type_>(                                   \
+            casted_array[(size_t)index[dim_zero]]);                            \
+    }                                                                          \
+                                                                               \
+    template <Dimensions TIndex>                                               \
+    static void set_element(native_type &array, const TIndex &index,           \
+                            element_type_ value) noexcept {                    \
+        static_assert(TIndex::rank() == 1, "index must be 1D");                \
+        auto casted_array = (cast_type)array;                                  \
+        casted_array[(size_t)index[dim_zero]] =                                \
+            std::bit_cast<bitcast_type>(value);                                \
+        array = (native_type)casted_array;                                     \
+    }
