+++ ntt/include/nncase/ntt/vector_ops.h
@@ -14,9 +14,11 @@
  */
 #pragma once
 #include "apply.h"
+#include "dimension.h"
 #include "primitive_ops.h"
 #include "tensor_traits.h"
 #include "vector.h"
+#include <type_traits>
 
 namespace nncase::ntt::ops {
 // unary_ops ops
@@ -541,6 +543,84 @@ template <Vector TVector1, Vector TVector2> struct cast<TVector1, TVector2> {
     ops::cast<from_type, to_type> op_;
 };
 
+template <Vector TFromVector, Scalar TTo> struct cast_elem<TFromVector, TTo> {
+    using TFromElem = typename TFromVector::element_type;
+
+    constexpr auto operator()(const TFromVector &froms) const noexcept
+        requires(sizeof(TFromElem) == sizeof(TTo))
+    {
+        if constexpr (std::is_same_v<TFromElem, TTo>) {
+            return froms; // No cast needed
+        } else {
+            using TToVector =
+                basic_vector<TTo, typename TFromVector::shape_type>;
+            return ntt::cast<TFromVector, TToVector>()(froms);
+        }
+    }
+
+    constexpr auto operator()(const TFromVector &froms) const noexcept
+        requires(sizeof(TFromElem) > sizeof(TTo))
+    {
+        if constexpr (TFromVector::rank() > 2) {
+            constexpr auto domain = TFromVector::shape().front();
+            using TToInnerVector =
+                std::remove_cv_t<decltype(ntt::cast_elem<TTo>(froms(0_dim)))>;
+            using to_shape_t =
+                std::remove_cv_t<decltype(TToInnerVector::shape().prepend(
+                    domain))>;
+
+            basic_vector<TTo, to_shape_t> tos;
+            ntt::loop<domain>([&](auto outer_index) {
+                tos(outer_index) = ntt::cast_elem<TTo>(froms(outer_index));
+            });
+            return tos;
+        } else {
+            constexpr auto N = TFromVector::shape().front();
+            static_assert(N == sizeof(TFromElem) / sizeof(TTo));
+            constexpr auto lanes = TFromVector::shape().back();
+
+            vector<TTo, N * lanes> tos;
+            ops::cast<TFromElem, TTo> cast_op;
+            ntt::loop<N>([&](auto n) {
+                ntt::loop<lanes>([&](auto lane) {
+                    tos(n * lanes + lane) = cast_op(froms(n, lane));
+                });
+            });
+            return tos;
+        }
+    }
+
+    constexpr auto operator()(const TFromVector &froms) const noexcept
+        requires(sizeof(TFromElem) < sizeof(TTo))
+    {
+        if constexpr (TFromVector::rank() > 1) {
+            constexpr auto domain = TFromVector::shape().front();
+            using TToInnerVector =
+                std::remove_cv_t<decltype(ntt::cast_elem<TTo>(froms(0_dim)))>;
+            using to_shape_t =
+                std::remove_cv_t<decltype(TToInnerVector::shape().prepend(
+                    domain))>;
+
+            basic_vector<TTo, to_shape_t> tos;
+            ntt::loop<domain>([&](auto outer_index) {
+                tos(outer_index) = ntt::cast_elem<TTo>(froms(outer_index));
+            });
+            return tos;
+        } else {
+            constexpr auto N = fixed_dim_v<sizeof(TTo) / sizeof(TFromElem)>;
+            constexpr auto lanes = TFromVector::shape().back() / N;
+
+            vector<TTo, N, lanes> tos;
+            ops::cast<TFromElem, TTo> cast_op;
+            ntt::loop<N>([&](auto n) {
+                ntt::loop<lanes>([&](auto lane) {
+                    tos(n, lane) = cast_op(froms(n * lanes + lane));
+                });
+            });
+            return tos;
+        }
+    }
+};
 } // namespace nncase::ntt::ops
 
 namespace nncase::ntt::vector_ops {
@@ -559,10 +639,9 @@ template <Vector TVector> struct vload_scalar {
     }
 };
 
-template <Vector TVector> struct vunaligned_load {
+template <Vector TVector, ScalarOrVector U> struct vunaligned_load {
     using T = typename TVector::element_type;
 
-    template <ScalarOrVector U>
     constexpr TVector operator()(const U *ptr) const noexcept {
         const auto domain =
             TVector::shape()
@@ -619,7 +698,7 @@ template <Scalar T, FixedShape Lanes>
 template <ScalarOrVector U>
 basic_vector<T, Lanes>
 basic_vector<T, Lanes>::unaligned_load_from(const U *ptr) noexcept {
-    return vector_ops::vunaligned_load<basic_vector<T, Lanes>>()(ptr);
+    return vector_ops::vunaligned_load<basic_vector<T, Lanes>, U>()(ptr);
 }
 
 template <bool AccC, bool TransA = false, Vector T1, Vector T2, Vector TResult>
